Add new feature trunk control and pom db-vs-local compare and show-catalina-log
 -trunk control : just show admin-user
 -pom compare: local pom.xml compare with savetodb for descoveer mis-prop
 -catalina.out: save tail -f catalina.out > db when restart-tomcat

+++ godzilla-service/src/main/java/cn/godzilla/service/impl/MvnProviderServiceImpl.java
@@ -50,7 +50,7 @@ public class MvnProviderServiceImpl extends Application implements MvnProviderSe
 			
 		}
 		//保存mvn部署日志
-		int logid = operateLogService.addOperateLog(mvnlog, jarlog);
+		Long logid = operateLogService.addOperateLog(mvnlog, jarlog);
 		
 		mvnBuildThreadLocal.set("FAILURE");
 		mvnERRORThreadLocal.set("SUCCESS");
